separate merkle_tree code out of merkle_set.rs
