+++ src/lib.rs
@@ -5,3 +5,4 @@ pub mod fast_forward;
 pub mod gen;
 pub mod generator_rom;
 pub mod merkle_set;
+pub mod merkle_tree;
